app/requirements.txt
flask
werkzeug
pillow
opencv-python-headless
numpy
tensorflow
gunicorn   # optional, for production